Frontend/requirements.txt
apache-airflow==2.5.3
streamlit
requests
pandas
snowflake-connector-python
python-dotenv
pyarrow>=10.0.1,<10.1.0
matplotlib
seaborn